36 Moved "taskLink" extension to its own file
